Added a beginning of access to the backend (throws exception)
+ changed Identifier to be BigIntiger instead of long

+++ Mezu/app/src/main/java/money/mezu/mezu/BudgetIdentifier.java
@@ -1,11 +1,13 @@
 package money.mezu.mezu;
 
+import java.math.BigInteger;
+
 /**
  * Created by asafb on 4/15/2017.
  */
 
 public class BudgetIdentifier extends Identifier {
-    public BudgetIdentifier(long id) {
+    public BudgetIdentifier(BigInteger id) {
         super(id);
     }
