+++ Mezu/app/src/main/java/money/mezu/mezu/UserIdentifier.java
@@ -1,11 +1,13 @@
 package money.mezu.mezu;
 
+import java.math.BigInteger;
+
 /**
  * Created by asafb on 4/15/2017.
  */
 
 public class UserIdentifier extends Identifier {
-    public UserIdentifier(long id) {
+    public UserIdentifier(BigInteger id) {
         super(id);
     }
